ネームスペースなどの調整
