+++ internalUsages/MemStat.Net.Web/MemoryInfo.cs
@@ -1,0 +1,1 @@
+internal record MemoryInfo(double TotalMemory, double MemoryUsagePercentage);
