+++ internalUsages/MemStat.Net.Web/Program.cs
@@ -59,10 +59,3 @@ app.MapGet(
         .UnwrapBox());
 
 app.Run();
-
-internal record MemoryInfo(double TotalMemory, double MemoryUsagePercentage);
-
-internal record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
-{
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-}
