+++ internalUsages/MemStat.Net.Web/WeatherForecast.cs
@@ -1,0 +1,4 @@
+internal record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
+{
+    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+}
