+++ src/Sekiban.Core/Documents/Pools/AggregateGroupStream.cs
@@ -1,0 +1,10 @@
+using Sekiban.Core.Aggregate;
+namespace Sekiban.Core.Documents.Pools;
+
+public record AggregateGroupStream(
+    string AggregateGroup,
+    AggregateContainerGroup AggregateContainerGroup = AggregateContainerGroup.Default) : IAggregatesStream
+{
+    public AggregateContainerGroup GetAggregateContainerGroup() => AggregateContainerGroup;
+    public List<string> GetStreamNames() => [AggregateGroup];
+}
