+++ src/Sekiban.Core/Documents/Pools/IPooledEventWriter.cs
@@ -1,0 +1,6 @@
+namespace Sekiban.Core.Documents.Pools;
+
+public interface IPooledEventWriter : IDocumentWriter
+{
+}
+
