+++ src/Sekiban.Infrastructure.Cosmos/Documents/CosmosDocumentRepository.cs
@@ -2,6 +2,7 @@ using Microsoft.Azure.Cosmos.Linq;
 using ResultBoxes;
 using Sekiban.Core.Aggregate;
 using Sekiban.Core.Documents;
+using Sekiban.Core.Documents.Pools;
 using Sekiban.Core.Documents.ValueObjects;
 using Sekiban.Core.Events;
 using Sekiban.Core.Exceptions;
